Added DevMode and Cleanup
Added a developer mode that enables a user to print the sent messages with a switch. Moved more hard coded settings to the settings file. Added a generic example vaccine_sites and false_positive_checks files. Also added the option to switch between texting and tweeting depending on the users needs.